% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_mapping.R
\name{analyze_column_mapping}
\alias{analyze_column_mapping}
\title{Analyze Individual Column Mapping}
\usage{
analyze_column_mapping(
  engine,
  source_col,
  target_col,
  source_data_col,
  target_sdmx
)
}
\description{
Analyze Individual Column Mapping
}
\keyword{internal}
