% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_transformations.R
\name{analyze_column_transformation}
\alias{analyze_column_transformation}
\title{Analyze Column Transformation}
\usage{
analyze_column_transformation(
  source_col,
  target_col,
  source_values,
  target_info,
  target_sdmx,
  llm_config
)
}
\description{
Analyze Column Transformation
}
\keyword{internal}
