% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_mapping.R
\name{apply_learning_boost}
\alias{apply_learning_boost}
\title{Apply Learning Boost}
\usage{
apply_learning_boost(engine, source_name, target_name)
}
\description{
Apply Learning Boost
}
\keyword{internal}
