% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_content_similarity}
\alias{calculate_content_similarity}
\title{Calculate content pattern similarity}
\usage{
calculate_content_similarity(source_col, target, data_analysis)
}
\description{
Calculate content pattern similarity
}
\keyword{internal}
