% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{calculate_data_quality_score}
\alias{calculate_data_quality_score}
\title{Calculate Data Quality Score}
\usage{
calculate_data_quality_score(source_profile)
}
\description{
Calculate Data Quality Score
}
\keyword{internal}
