% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_structure.R
\name{calculate_mapping_complexity}
\alias{calculate_mapping_complexity}
\title{Calculate mapping complexity score}
\usage{
calculate_mapping_complexity(dim_comp, attr_comp, meas_comp)
}
\description{
Calculate mapping complexity score
}
\keyword{internal}
