% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_type_compatibility}
\alias{calculate_type_compatibility}
\title{Calculate type compatibility}
\usage{
calculate_type_compatibility(source_col, target)
}
\description{
Calculate type compatibility
}
\keyword{internal}
