% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{check_hierarchical_nesting}
\alias{check_hierarchical_nesting}
\title{Check Hierarchical Nesting}
\usage{
check_hierarchical_nesting(parent_col, child_col)
}
\description{
Check Hierarchical Nesting
}
\keyword{internal}
