% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Validate that required packages are available}
\usage{
check_packages(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\description{
Validate that required packages are available
}
\keyword{internal}
