% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_default_workflow_steps}
\alias{create_default_workflow_steps}
\title{Create Default Workflow Steps}
\usage{
create_default_workflow_steps()
}
\description{
Create Default Workflow Steps
}
\keyword{internal}
