% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_transformations.R
\name{create_transformation_context}
\alias{create_transformation_context}
\title{Create Transformation Context}
\usage{
create_transformation_context(
  transformation_suggestions,
  source_data,
  target_sdmx
)
}
\description{
Create Transformation Context
}
\keyword{internal}
