% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_workflow_recommendations}
\alias{create_workflow_recommendations}
\title{Create Workflow Recommendations}
\usage{
create_workflow_recommendations(workflow)
}
\description{
Create Workflow Recommendations
}
\keyword{internal}
