% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_mapping.R
\name{generate_mapping_recommendations}
\alias{generate_mapping_recommendations}
\title{Generate Mapping Recommendations}
\usage{
generate_mapping_recommendations(
  mappings,
  coverage_analysis,
  source_profile,
  target_sdmx
)
}
\description{
Generate Mapping Recommendations
}
\keyword{internal}
