% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{generate_mapping_report}
\alias{generate_mapping_report}
\title{Generate Mapping Report}
\usage{
generate_mapping_report(mapping_result, source_profile)
}
\description{
Generate Mapping Report
}
\keyword{internal}
