% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_transformations.R
\name{generate_recoding_code}
\alias{generate_recoding_code}
\title{Generate Recoding Code}
\usage{
generate_recoding_code(source_col, target_col, recoding_analysis)
}
\description{
Generate Recoding Code
}
\keyword{internal}
