% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_mapping.R, R/value_transformations.R
\name{get_target_column_info}
\alias{get_target_column_info}
\title{Get Target Column Information}
\usage{
get_target_column_info(target_col, target_sdmx)

get_target_column_info(target_col, target_sdmx)
}
\description{
Get Target Column Information

Get Enhanced Target Column Information
}
\keyword{internal}
