% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_providers.R
\docType{data}
\name{LLM_PROVIDERS}
\alias{LLM_PROVIDERS}
\title{LLM Provider Types}
\format{
An object of class \code{list} of length 8.
}
\usage{
LLM_PROVIDERS
}
\description{
LLM Provider Types
}
\keyword{datasets}
