% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_mapping.R
\docType{data}
\name{MAPPING_CONFIDENCE}
\alias{MAPPING_CONFIDENCE}
\title{Mapping Confidence Levels}
\format{
An object of class \code{list} of length 5.
}
\usage{
MAPPING_CONFIDENCE
}
\description{
Confidence levels for mapping suggestions.
}
\keyword{datasets}
