% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{print.llmx_data_analysis}
\alias{print.llmx_data_analysis}
\title{Print method for data analysis results}
\usage{
\method{print}{llmx_data_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class "llmx_data_analysis"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for data analysis results
}
