% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_providers.R
\name{print.llmx_llm_config}
\alias{print.llmx_llm_config}
\title{Print method for LLM config}
\usage{
\method{print}{llmx_llm_config}(x, ...)
}
\arguments{
\item{x}{An object of class "llmx_llm_config"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for LLM config
}
