% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.llmx_validation_result}
\alias{print.llmx_validation_result}
\title{Print method for validation results}
\usage{
\method{print}{llmx_validation_result}(x, ...)
}
\arguments{
\item{x}{An object of class "llmx_validation_result"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for validation results
}
