% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{print.llmx_workflow_result}
\alias{print.llmx_workflow_result}
\title{Print method for workflow results}
\usage{
\method{print}{llmx_workflow_result}(x, ...)
}
\arguments{
\item{x}{An object of class "llmx_workflow_result"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for workflow results
}
