% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_transformations.R
\name{suggest_derived_transformations}
\alias{suggest_derived_transformations}
\title{Suggest Derived Transformations}
\usage{
suggest_derived_transformations(mapping_result, source_data, target_sdmx)
}
\description{
Suggest Derived Transformations
}
\keyword{internal}
