% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate overall data structure}
\usage{
validate_data_structure(data, sdmx_metadata)
}
\description{
Validate overall data structure
}
\keyword{internal}
