% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dimension_values}
\alias{validate_dimension_values}
\title{Validate dimension values against codelists}
\usage{
validate_dimension_values(data, sdmx_metadata)
}
\description{
Validate dimension values against codelists
}
\keyword{internal}
