% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{data}
\name{WORKFLOW_STATUS}
\alias{WORKFLOW_STATUS}
\title{Workflow Step Status}
\format{
An object of class \code{list} of length 5.
}
\usage{
WORKFLOW_STATUS
}
\description{
Workflow Step Status
}
\keyword{datasets}
